ajout getLivreATitre(); getClassification() dans actionbd, ajout de test dans testbd; raphe

+++ src/main/Livre.java
@@ -2,13 +2,13 @@ package main;
 
 public class Livre {
 
-    private int isbn;
+    private long isbn;
     private String titre;
     private int nbpages;
     private String datepubli;
     private double prix;
 
-    public Livre(int isbn,String titre, int nbpages, String datepubli, double prix){
+    public Livre(long isbn,String titre, int nbpages, String datepubli, double prix){
         this.isbn=isbn;
         this.titre=titre;
         this.nbpages=nbpages;
@@ -17,7 +17,7 @@ public class Livre {
     }
     
     public double getPrix(){return this.prix;}
-    public int getISBN(){return this.isbn;}
+    public long getISBN(){return this.isbn;}
     public int getNbpages(){return this.nbpages;}
     public String getTitre(){return this.titre;}
     public String getDatepubli(){return this.datepubli;}
